+++ lib/Effects/Effects/HueSpin.h
@@ -1,15 +1,15 @@
-void hueSpinEffectLoop(struct EffectState *effectState, unsigned int loopMs)
+void hueSpinEffectLoop(struct EffectState *effectState, uint8_t loopMs)
 {
-  const unsigned int loopPosition = effectState->loopPosition % loopMs;
+  const uint8_t loopPosition = effectState->loopPosition % loopMs;
   const float loopPercent = float(loopPosition) / loopMs;
 
   const int loopOffset = loopPercent * config.LED_CNT;
 
-  for (int i = 0; i < config.LED_CNT; i++)
+  for (uint16_t i = 0; i < config.LED_CNT; i++)
   {
     const float ledPercent = float(i) / config.LED_CNT;
-    const unsigned int hueOffset = ledPercent * 255;
-    const unsigned int idx = (i + loopOffset) % config.LED_CNT;
+    const uint8_t hueOffset = ledPercent * 255;
+    const uint8_t idx = (i + loopOffset) % config.LED_CNT;
     leds[idx] = CHSV(hueOffset, 255, 255);
   }
