+++ lib/Effects/Effects/PatternLoop.h
@@ -1,19 +1,19 @@
 
-void patternLoopEffectLoop(struct EffectState *effectState, unsigned int stepDelays[], const CHSV stepColors[], unsigned int stepCnt)
+void patternLoopEffectLoop(struct EffectState *effectState, uint8_t stepDelays[], const CHSV stepColors[], uint8_t stepCnt)
 {
-  unsigned int loopMs = 0;
-  unsigned int step = 0;
+  uint8_t loopMs = 0;
+  uint8_t step = 0;
 
-  for (unsigned int i = 0; i < stepCnt; i++)
+  for (uint8_t i = 0; i < stepCnt; i++)
   {
     loopMs += stepDelays[i];
   }
 
-  const unsigned int loopPosition = effectState->loopPosition % loopMs;
+  const uint8_t loopPosition = effectState->loopPosition % loopMs;
 
-  unsigned int loopCnt = 0;
+  uint8_t loopCnt = 0;
 
-  for (unsigned int i = 0; i < stepCnt; i++)
+  for (uint8_t i = 0; i < stepCnt; i++)
   {
     loopCnt += stepDelays[i];
     if (loopPosition > loopCnt)
